Move Home() and TiviContent() to :ui:root
